code improvements MealScreen
